+++ src/template.html
@@ -17,17 +17,16 @@
           <i class="fa fa-refresh fa-2x"></i>
         </div>
         <div class="to-do-add">
-          <form class="form">
-            <input
-              type="text"
-              class="to-do-add-text"
-              placeholder="Add to your list..."
-            />
-          </form>
-          <i class="fas fa-rotate-90 enter-btn">&#xf3be;</i>
+          <input
+            type="text"
+            class="to-do-add-text"
+            placeholder="Add to your list..."
+          />
+          <i class="fas fa-rotate-90">&#xf3be;</i>
         </div>
-
-        <ul class="to-do-list"></ul>
+        <div class="to-do-list-container">
+          <ul class="to-do-list"></ul>
+        </div>
       </div>
       <button class="to-do-clear">Clear all completed</button>
     </div>
